deletion of routes folder
