--- routes/api/index.js
@@ -1,11 +1,0 @@
-const router = require('express').Router();
-
-const userRoutes = require('./user-routes');
-const postRoutes = require('./post-routes');
-const commentRoutes = require('./comment-routes');
-
-router.use('/user', userRoutes);
-router.use('/post', postRoutes);
-router.use('/comment', commentRoutes);
-
-module.exports = router;
