--- routes/api/post-routes.js
@@ -1,72 +1,0 @@
-const router = require('express').Router();
-const { User, Posts, Comments } = require('../../models');
-const withAuth = require('../../utils/auth');
-
-// this is going to create a new post
-
-router.post("/", withAuth, async (req, res) => {
-  try {
-    const newPost = await Posts.create({
-      ...req.body,
-      userId: req.session.userId,
-    });
-    console.log("This is the new post", newPost);
-    res.status(200).json(newPost);
-  } catch (err) {
-    res.status(400).json(err);
-  }
-});
-
-// this is going to be able to edit a post
-
-router.put("/:id", withAuth, async (req, res) => {
-  try {
-    const updatedPost = await Posts.update(
-      {
-        title: req.body.title,
-        content: req.body.content,
-      },
-      {
-        where: {
-          id: req.params.id,
-        },
-      }
-    );
-    if (!updatedPost) {
-      res.status(404).json({ message: "This id has no post" });
-      return;
-    }
-    res.status(200).json(updatedPost);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-// this is going to be able to delete a post
-
-router.delete("/:id", withAuth, async (req, res) => {
-  try {
-    const commentData = await Comments.destroy({
-      where: { postId: req.params.id },
-    });
-
-    const postData = await Posts.destroy({
-      where: {
-        id: req.params.id,
-        userId: req.session.userId,
-      },
-    });
-    if (!postData) {
-      res.status(404).json({
-        message: `No user Id ${req.session.userId} found with id = ${req.params.id}`,
-      });
-      return;
-    }
-
-    res.status(200).json(postData);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-module.exports = router;
